refactor: reorganize project structure by separating database and timer modules into distinct files
